[Patchset 2] trace_processor: remove almost all usages of uint64_t
The use of uint64_t is quite dangerous as SQLite cannot actually represent
it when reporting it tot users. Therefore, using it on any type which is
intended for displaying to users is highly misleading and should not
be used.

Therefore, switch over all types (with the exception of stack ids) to
int64_t or uint32_t. Stack ids are ok to stay as they are as they are
a hash of the id and no one should be relying on their values (it
can change over runs of the trace processor and is only required to be
internally consistent).

Bug: 120278869

+++ src/trace_processor/slice_tracker.cc
@@ -72,8 +72,7 @@ void SliceTracker::StartSlice(int64_t timestamp,
     PERFETTO_DFATAL("Slices with too large depth found.");
     return;
   }
-  uint64_t parent_stack_id =
-      depth == 0 ? 0 : slices->stack_ids()[stack->back()];
+  int64_t parent_stack_id = depth == 0 ? 0 : slices->stack_ids()[stack->back()];
   size_t slice_idx = slices->AddSlice(timestamp, duration, utid, cat, name,
                                       depth, 0, parent_stack_id);
   stack->emplace_back(slice_idx);
@@ -147,7 +146,7 @@ void SliceTracker::MaybeCloseStack(int64_t ts, SlicesStack* stack) {
   }
 }
 
-uint64_t SliceTracker::GetStackHash(const SlicesStack& stack) {
+int64_t SliceTracker::GetStackHash(const SlicesStack& stack) {
   PERFETTO_DCHECK(!stack.empty());
 
   const auto& slices = context_->storage->nestable_slices();
@@ -162,7 +161,7 @@ uint64_t SliceTracker::GetStackHash(const SlicesStack& stack) {
              sizeof(slices.names()[slice_idx]));
   }
   constexpr uint64_t kMask = uint64_t(-1) >> 1;
-  return (std::hash<std::string>{}(s)) & kMask;
+  return static_cast<int64_t>((std::hash<std::string>{}(s)) & kMask);
 }
 
 }  // namespace trace_processor
